Add files via upload
This is a JEE project for a particular hospital . In this patients can login and sign up to their account,can book for their appointments with particular doctor and can even book their prescribed medicines by the doctor. Since the database is not attached , the application will not run, only the jsp and servlet pages are available.